refactor(football): extract services for cleaner code
- score_formatter.py: match formatting logic
- player_assets.py: player photo/badge fetching

Reduced LOC:
- score.py: 204 -> ~90
- football_team.py: 244 -> ~195

All tests pass (1814).
